app/requirements.txt
alembic==1.16.3
asyncpg==0.30.0
bcrypt==4.3.0
black==25.1.0
fastapi==0.116.0
flake8==7.3.0
greenlet==3.2.3
httpx==0.28.1
isort==6.0.1
psycopg==3.2.9
pydantic[email]==2.11.7
pydantic-settings==2.10.1
pyjwt[crypto]==2.10.1
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-cov==6.2.1
python-multipart==0.0.20
uvicorn==0.35.0